NF: move CheckDatabase where it's used
