+++ AnkiDroid/src/main/java/com/ichi2/async/CollectionTask.java
@@ -1165,28 +1165,6 @@ public class CollectionTask<Progress, Result> extends BaseAsyncTask<Void, Progre
         }
     }
 
-    public static class CheckDatabase implements TaskDelegate<String, Pair<Boolean, Collection.CheckDatabaseResult>> {
-        public Pair<Boolean, Collection.CheckDatabaseResult> task(@NonNull Collection col, @NonNull ProgressSenderAndCancelListener<String> collectionTask) {
-            Timber.d("doInBackgroundCheckDatabase");
-            // Don't proceed if collection closed
-            if (col == null) {
-                Timber.e("doInBackgroundCheckDatabase :: supplied collection was null");
-                return new Pair<>(false, null);
-            }
-            
-            Collection.CheckDatabaseResult result = col.fixIntegrity(new TaskManager.ProgressCallback(collectionTask, AnkiDroidApp.getAppResources()));
-            if (result.getFailed()) {
-                //we can fail due to a locked database, which requires knowledge of the failure.
-                return new Pair<>(false, result);
-            } else {
-                // Close the collection and we restart the app to reload
-                CollectionHelper.getInstance().closeCollection(true, "Check Database Completed");
-                return new Pair<>(true, result);
-            }
-        }
-    }
-
-
     public static class RebuildCram implements TaskDelegate<Void, StudyOptionsFragment.DeckStudyData> {
         public StudyOptionsFragment.DeckStudyData task(@NonNull Collection col, @NonNull ProgressSenderAndCancelListener<Void> collectionTask) {
             Timber.d("doInBackgroundRebuildCram");
